Items are no longer processed multiple times
Phabulous sometimes reacted to messages from itself. That resulted in
recursive lookups. Some messages that phabulous posts, e.g. via the
feed, are not posted as a user. For example when the config
`slack.as-user` is set to `false`. In that case phabulous did not
realize that it sent the message itself and processed all mentions. Now,
phabulous only processes messages that have a user. The downside is,
that it no longer reacts to some messages from other bots.

Phabulous did not correctly track the items it already processed in a
single message. It now checks correctly that every match is processed
only once.